chore: refactor tap to qr and separate shared logic
