+++ addon/popup/tap-to-qr.html
@@ -26,6 +26,6 @@
     </div>
     <script type="application/javascript" src="../libs/browser-polyfill.js" defer></script>
     <script src="bundle.js" defer></script>
-    <script src="tap-to-qr.js" defer></script>
+    <script type="module" src="tap-to-qr.js" defer></script>
 </body>
 </html>
